Update README.md
My favorite Series was an Anime, it was all about a child born with hatred from the people around him. The series was all about a person living from misery but still achieving his dreams despite many obstacles that was present in the path he chooses. The anime series was NARUTO, it was my childhood series and up until now I still love the storyline it has.